Fix and refactor validation of omit authors
It's better off to treat it as a custom argparse type instead of running
a separate function.

Also there was a bug previously where if you omit no authors it would
still say there was 1 omit author. This didn't affect the list of
winners but it's still a good idea to report 0 when there's no authors
being omit.