Simplify output into 1 big f string
This output was slowly built up which is why I didn't think to make it
an f string from the start.

Also using this strategy allows us to easily have 2 spaces before the
output of the winners. This follows the pattern for the verbose output
of displaying the comment authors' disply names.

This has a benefit in that if you were to pipe the output of this
script into grep or some other tool you could easily filter out lines
that start with 2 spaces to get a list of winners.